Image selection implemented
Home page design, exercise list multi-selection, video feature and note
taking feature 100%
- automatic instructions in notes
- note cell position
- note cell editing mode bug to fix
- home page selector arrow view position
- Image customization in launcher

+++ BarcodeFitness-iPhone/BFExerciseListViewController.m
@@ -133,7 +133,7 @@
     if (tableView == self.searchDisplayController.searchResultsTableView) {
         eIndex = currentExercise.row;
     } else {
-        eIndex = indexPath.row;
+        eIndex = (int)indexPath.row;
     }
     if ([[_wishList objectAtIndex:eIndex] isEqualToString:@"YES"]) {
         cell.accessoryType = UITableViewCellAccessoryCheckmark;
